+++ test/angular_material_tests.dart
@@ -1,0 +1,55 @@
+// Copyright (C) 2015 Sergey Akopkokhyants.
+// https://github.com/akserg/angular.dart.material
+// All rights reserved.  Please see the LICENSE.md file.
+library angular_material_test;
+
+import 'package:guinness/guinness_html.dart';
+import 'dart:async';
+import 'dart:html' as dom;
+
+import 'package:angular/angular.dart';
+import 'package:angular/core/module_internal.dart';
+import 'package:angular/mock/module.dart';
+import 'package:angular/mock/test_injection.dart';
+
+import 'package:angular_material/material.dart';
+
+part 'unit/ripple_test.dart';
+part 'unit/material_test.dart';
+
+main(){
+  guinnessEnableHtmlMatchers();
+
+  testRipple();
+  testMaterial();
+  
+  guinness.initSpecs();
+}
+
+loadTemplates(List<String> templates){
+  updateCache(template, response) => inject((TemplateCache cache) => cache.put(template, response));
+
+  final futures = templates.map((template) =>
+    dom.HttpRequest.request('packages/angular_material/' + template, method: "GET").
+    then((_) => updateCache(template, new HttpResponse(200, _.response))));
+
+  return Future.wait(futures);
+}
+
+compileComponent(String html, Map scope, callback){
+  return async(() {
+    inject((TestBed tb) {
+      final s = tb.rootScope.createChild(scope);
+      final el = tb.compile('<div>$html</div>', scope: s);
+
+      microLeap();
+      digest();
+
+      callback(s, el);
+    });
+  });
+}
+
+digest(){
+  inject((TestBed tb) { tb.rootScope.apply(); });
+}